requirements.txt
streamlit
pandas
mysql-connector-python
python-dotenv
unidecode
